refactor(beta): unify Agent/LiveAgent/Plugin via PluginTarget hierarchy
Extract the shared authoring and construction surface of Agent, LiveAgent,
and Plugin into two base classes in autogen/beta/plugin.py:

- PromptObserverMixin: prompt/observer/tool decorators, the
  add_middleware/insert_middleware/add_policy/add_observer collectors, and
  the abstract add_tool hook — shared by all three.
- PluginTarget(PromptObserverMixin): the runnable targets' shared
  constructor (_init_target), eager add_tool, wrap-now hitl_hook,
  _tool_executor, and _apply_plugin.

Agent and LiveAgent now inherit PluginTarget; Plugin stays on the bare
mixin as a pure contribution source. Remove Plugin.register /
NetworkPlugin.register — a plugin collects every contribution (tools,
middleware, observers, prompts, hooks, assembly policies) and the single
_apply_plugin copies them onto the target. NetworkPlugin collects its
context policy in __init__; HubClient applies it via agent._apply_plugin.

Along the way: fix a latent bug where Agent.add_tool delegated to the
abstract mixin hook (NotImplementedError), and restore plugin support on
LiveAgent (it lacked _apply_plugin).

See docs/adr/0001-plugin-target-hierarchy.md.
